Add directory. See #26.
